refactoring of profile page and ui completed
